+++ examples/webview_example_native_ios/Pods/Target Support Files/react-native-branch/react-native-branch-umbrella.h
@@ -10,6 +10,8 @@
 #endif
 #endif
 
+#import "BranchContentMetadata+RNBranch.h"
+#import "BranchEvent+RNBranch.h"
 #import "BranchLinkProperties+RNBranch.h"
 #import "BranchUniversalObject+RNBranch.h"
 #import "NSObject+RNBranch.h"